Added functionality to register player data
